+++ highwave/src/com/dudko/highwave/deposit/deposits/CatchMomentDeposit.java
@@ -17,8 +17,8 @@ public abstract class CatchMomentDeposit extends Deposit {
 	}
 
 	@Override
-	public DepositAccount calculateDeposit(float amount, int period) {
-		float minOpenAmount = 1000.0f;
+	public DepositAccount calculateDeposit(long amount, int period) {
+		long minOpenAmount = 1000;
 		if (amount < minOpenAmount) {
 			return null;
 		}
