+++ highwave/src/com/dudko/highwave/deposit/deposits/PrudentDeposit.java
@@ -21,8 +21,8 @@ public class PrudentDeposit extends Deposit {
 	}
 
 	@Override
-	public DepositAccount calculateDeposit(float amount, int period) {
-		float minOpenAmount = 100.0f;
+	public DepositAccount calculateDeposit(long amount, int period) {
+		long minOpenAmount = 100;
 		if (amount < minOpenAmount) {
 			return null;
 		}
